Rewrote #3
Episode #3 is rewritten! This one was a little more work than the
previous one, but a lot of fun! The Vanilla Js version actually makes
more sense to me, but that might be because I wrote it myself. Also, the
previous commit was needed to fix a merge conflict. Also, longest commit
message ever!